+++ lib/dusk/native.ex
@@ -17,4 +17,6 @@ defmodule Dusk.Native do
 
   def key_match(_pat, _key), do: :erlang.nif_error(:nif_not_loaded)
   def hash64(_bin), do: :erlang.nif_error(:nif_not_loaded)
+  def blake3(_bin), do: :erlang.nif_error(:nif_not_loaded)
+  def xxh3(_bin), do: :erlang.nif_error(:nif_not_loaded)
 end